Replace GraphHopper with OpenRouteService for route calculations
Refactored route calculation logic to use OpenRouteService instead of GraphHopper. Introduced `utils/get_route.py` for modular handling of route calculations and removed GraphHopper dependencies from `core.py`. Adjusted related function calls and configuration accordingly.
