+++ core.py
@@ -1,11 +1,10 @@
-from opencage.geocoder import OpenCageGeocode
 import streamlit as st
 import pandas as pd
-import json
 import base64
 from github import Github, GithubException
 from io import BytesIO
 from streamlit_pdf_viewer import pdf_viewer
+import json
 
 @st.cache_data
 def charger_donnees(nom_fichier="data/hebergements_chemins.parquet", format=None, branche="main"):
@@ -144,249 +143,6 @@ def sauvegarder_donnees(contenu, nom_fichier, message_commit="Mise à jour des d
         st.error(traceback.format_exc())
         return False
 
-
-def add_lat_lon(df, address_column="Adresse"):
-    """Ajoute les coordonnées géographiques (latitude, longitude) pour chaque adresse"""
-    try:
-        api_key = st.secrets["opencage"]["api_key"]
-    except KeyError:
-        st.error("Clé API OpenCage manquante. Vérifiez le fichier .streamlit/secrets.toml.")
-        return df
-
-    geocoder = OpenCageGeocode(api_key)
-
-    # Créer les colonnes Latitude et Longitude si elles n'existent pas
-    if "Latitude" not in df.columns:
-        df["Latitude"] = None
-    if "Longitude" not in df.columns:
-        df["Longitude"] = None
-
-    def get_coordinates(address):
-        try:
-            result = geocoder.geocode(address)
-            if result:
-                return result[0]["geometry"]["lat"], result[0]["geometry"]["lng"]
-        except Exception as e:
-            st.error(f"Erreur pour {address} : {e}")
-        return None, None
-
-    for index, row in df.iterrows():
-        if pd.isna(row["Latitude"]) or pd.isna(row["Longitude"]):
-            lat, lon = get_coordinates(row[address_column])
-            df.at[index, "Latitude"] = lat
-            df.at[index, "Longitude"] = lon
-
-    st.info("✅ Latitude et Longitude ajoutées avec succès !")
-    return df
-
-
-def get_graphhopper_route(start_coords, end_coords, type_deplacement="Marche"):
-    """
-    Calcule un itinéraire en utilisant l'API GraphHopper.
-
-    Args:
-        start_coords: Tuple (latitude, longitude) du point de départ
-        end_coords: Tuple (latitude, longitude) du point d'arrivée
-        type_deplacement: Type de déplacement ('Marche' ou 'Voiture')
-
-    Returns:
-        Un tuple (distance_km, duration_hours, coords_json) ou (None, None, None) en cas d'erreur
-    """
-    import requests
-    import json
-
-    # Vérifier que les coordonnées sont valides
-    if None in (start_coords[0], start_coords[1], end_coords[0], end_coords[1]):
-        print("Coordonnées invalides, impossible de calculer l'itinéraire.")
-        return None, None, None
-
-    # Déterminer le profil GraphHopper en fonction du type de déplacement
-    if not isinstance(type_deplacement, str):
-        print(f"Type de déplacement invalide: {type_deplacement} (type: {type(type_deplacement)})")
-        return None, None, None
-
-    if type_deplacement == "Marche":
-        profile = "foot"  # Utiliser "foot" au lieu de "hike" car "hike" n'est pas standard
-    elif type_deplacement == "Voiture":
-        profile = "car"
-    else:
-        print(f"Type de déplacement non reconnu: {type_deplacement}")
-        return None, None, None
-
-    # Ta clé API GraphHopper
-    api_key = st.secrets["graphhopper"]["token"]
-
-    # Construire l'URL de l'API GraphHopper
-    base_url = "https://graphhopper.com/api/1/route"
-
-    # Formatage correct des paramètres selon la documentation
-    params = {
-        "key": api_key,
-        "profile": profile,  # Utiliser "profile" au lieu de "vehicle"
-        "points_encoded": "false",
-        "instructions": "false",
-        "calc_points": "true",
-        "locale": "fr"
-    }
-
-    # Ajouter les points de manière correcte (séparément pour chaque point)
-    params["point"] = [
-        f"{start_coords[0]},{start_coords[1]}",
-        f"{end_coords[0]},{end_coords[1]}"
-    ]
-
-    try:
-        response = requests.get(base_url, params=params)
-
-        # Imprimer l'URL complète pour le débogage
-        print(f"URL de la requête: {response.url}")
-
-        response.raise_for_status()  # Lève une exception si le statut n'est pas 2xx
-
-        data = response.json()
-
-        if "paths" in data and len(data["paths"]) > 0:
-            path = data["paths"][0]
-
-            # Extraire les informations pertinentes
-            distance_km = path["distance"] / 1000  # Conversion en km
-            duration_hours = path["time"] / 3600000  # Conversion de ms en heures
-
-            # Extraire les coordonnées du chemin
-            route_coords = []
-            for point in path["points"]["coordinates"]:
-                # GraphHopper retourne les coordonnées dans l'ordre [longitude, latitude, élévation]
-                # Nous voulons [latitude, longitude] pour être compatible avec le reste du code
-                route_coords.append([point[1], point[0]])
-
-            # Convertir en JSON pour stockage
-            route_coords_json = json.dumps(route_coords)
-
-            return distance_km, duration_hours, route_coords_json
-        else:
-            print("Aucun itinéraire trouvé dans la réponse GraphHopper")
-            return None, None, None
-
-    except requests.exceptions.RequestException as e:
-        print(f"Erreur lors de la requête GraphHopper: {e}")
-        print(f"Réponse: {response.text if 'response' in locals() else 'N/A'}")
-        return None, None, None
-    except (KeyError, IndexError, ValueError) as e:
-        print(f"Erreur lors du traitement de la réponse GraphHopper: {e}")
-        return None, None, None
-
-
-def calculate_routes_graphhopper(df):
-    """Calcule les distances, durées et les trajets avec GraphHopper si non enregistrés."""
-
-
-    # Créer une copie du DataFrame pour éviter de modifier l'original
-    df = df.copy()
-
-    # Vérifier s'il y a des coordonnées manquantes et les ajouter
-    missing_coords = df[df["Latitude"].isna() | df["Longitude"].isna()].index
-    if len(missing_coords) > 0:
-        # Appliquer add_lat_lon seulement aux lignes avec coordonnées manquantes
-        df_missing = df.loc[missing_coords].copy()
-        df_missing = add_lat_lon(df_missing)
-
-        # Mettre à jour le DataFrame original avec les nouvelles coordonnées
-        for idx in missing_coords:
-            df.loc[idx, "Latitude"] = df_missing.loc[idx, "Latitude"]
-            df.loc[idx, "Longitude"] = df_missing.loc[idx, "Longitude"]
-
-    # Initialiser les listes pour stocker les résultats
-    distances = []
-    durations = []
-    route_geoms = []
-
-    # Variable pour suivre les itinéraires déjà calculés entre des paires de points
-    routes_cache = {}
-
-    # Calculer les itinéraires pour chaque segment
-    for i in range(len(df) - 1):  # On parcourt jusqu'à l'avant-dernier point
-        lat1, lon1 = df.iloc[i]["Latitude"], df.iloc[i]["Longitude"]
-        lat2, lon2 = df.iloc[i + 1]["Latitude"], df.iloc[i + 1]["Longitude"]
-
-        # Récupérer le type de déplacement
-        if "Type_Deplacement" in df.columns and pd.notna(df.iloc[i]["Type_Deplacement"]):
-            type_deplacement = df.iloc[i]["Type_Deplacement"]
-        else:
-            print(f"Type de déplacement non spécifié pour le segment {i} à {i + 1}")
-            type_deplacement = None
-
-        # Vérifier si toutes les coordonnées sont valides
-        valid_coords = pd.notna(lat1) and pd.notna(lon1) and pd.notna(lat2) and pd.notna(lon2)
-
-        # Créer une clé unique pour cette paire de coordonnées et type de déplacement
-        if valid_coords and type_deplacement is not None:
-            route_key = f"{lat1},{lon1}|{lat2},{lon2}|{type_deplacement}"
-
-            # Vérifier si l'itinéraire est déjà dans le cache
-            if route_key in routes_cache:
-                distance, duration, route_coords = routes_cache[route_key]
-            # Si déjà calculé dans le DataFrame et type de déplacement identique, on utilise cette valeur
-            elif pd.notna(df.iloc[i]["Chemin"]) and pd.notna(df.iloc[i]["Distance (km)"]):
-                try:
-                    route_coords = df.iloc[i]["Chemin"]
-                    if isinstance(route_coords, str):
-                        route_coords = json.loads(route_coords)
-                    distance = df.iloc[i]["Distance (km)"]
-                    duration = df.iloc[i]["Durée (h)"] if pd.notna(df.iloc[i]["Durée (h)"]) else None
-
-                    # Si la durée n'est pas disponible, on calcule l'itinéraire
-                    if duration is None:
-                        start_coords = (lat1, lon1)
-                        end_coords = (lat2, lon2)
-                        distance, duration, route_coords = get_graphhopper_route(start_coords, end_coords,
-                                                                                 type_deplacement)
-                        # Mettre en cache
-                        routes_cache[route_key] = (distance, duration, route_coords)
-                except Exception as e:
-                    print(f"Erreur lors de la lecture du chemin à l'index {i}: {e}")
-                    start_coords = (lat1, lon1)
-                    end_coords = (lat2, lon2)
-                    distance, duration, route_coords = get_graphhopper_route(start_coords, end_coords, type_deplacement)
-                    # Mettre en cache
-                    routes_cache[route_key] = (distance, duration, route_coords)
-            else:
-                # Sinon on calcule un nouveau tracé
-                start_coords = (lat1, lon1)
-                end_coords = (lat2, lon2)
-                distance, duration, route_coords = get_graphhopper_route(start_coords, end_coords, type_deplacement)
-                # Mettre en cache
-                routes_cache[route_key] = (distance, duration, route_coords)
-        else:
-            # Coordonnées invalides ou type de déplacement non défini
-            if not valid_coords:
-                print(f"Coordonnées manquantes pour le segment {i} à {i + 1}, impossible de calculer l'itinéraire.")
-            else:
-                print(
-                    f"Type de déplacement non défini pour le segment {i} à {i + 1}, impossible de calculer l'itinéraire.")
-            distance = None
-            duration = None
-            route_coords = json.dumps([])
-
-        # Mettre à jour le DataFrame directement
-        df.at[i, "Distance (km)"] = distance
-        df.at[i, "Durée (h)"] = duration
-
-        # S'assurer que route_coords est au format JSON
-        if isinstance(route_coords, list):
-            route_coords = json.dumps(route_coords)
-        df.at[i, "Chemin"] = route_coords
-
-        # Stocker pour retour de fonction
-        distances.append(distance)
-        durations.append(duration)
-        route_geoms.append(route_coords)
-
-    # Ajouter une dernière valeur pour correspondre à la taille du DataFrame
-    distances.append(None)
-    durations.append(None)
-    route_geoms.append(json.dumps([]))
-
-    return distances, durations, route_geoms, df
 
 @st.cache_data()
 def identifier_sejours_multiples(df):
